+++ apps/app-mosh-lol/src/pages/_document.tsx
@@ -1,0 +1,3 @@
+import { DocumentRoot } from "@sentrei/roots";
+
+export default DocumentRoot;