+++ apps/www-metascore-wtf/src/pages/_app.tsx
@@ -1,0 +1,28 @@
+import "tailwindcss/tailwind.css";
+import { SeoRoot, AnalyticsRoot } from "@sentrei/roots";
+import { ThemeProvider } from "next-themes";
+import type { AppProps } from "next/app";
+import type { FC } from "react";
+import { RecoilRoot } from "recoil";
+
+const AppRoot: FC<AppProps> = ({ Component, pageProps }) => {
+  return (
+    <>
+      <AnalyticsRoot />
+      <SeoRoot
+        title="Metascore"
+        subTitle="Metascore"
+        description="Metascore"
+        url="https://metascore.wtf"
+        twitter="@metascore"
+      />
+      <RecoilRoot>
+        <ThemeProvider attribute="class" defaultTheme="system">
+          <Component {...pageProps} />
+        </ThemeProvider>
+      </RecoilRoot>
+    </>
+  );
+};
+
+export default AppRoot;