+++ apps/www-mosh-lol/CHANGELOG.md
@@ -1,3 +1,21 @@
+## @sentrei/www-mosh-lol [1.6.1](https://github.com/sentrei/sentrei/compare/@sentrei/www-mosh-lol@1.6.0...@sentrei/www-mosh-lol@1.6.1) (2021-09-23)
+
+### Performance Improvements
+
+- ini tsc e2d ([299f23e](https://github.com/sentrei/sentrei/commit/299f23e4bc09c199ec375ac894f3e8d6709a94be))
+
+# @sentrei/www-mosh-lol [1.6.0](https://github.com/sentrei/sentrei/compare/@sentrei/www-mosh-lol@1.5.1...@sentrei/www-mosh-lol@1.6.0) (2021-09-23)
+
+### Features
+
+- ini app mosh lol with json sort ([3bd1255](https://github.com/sentrei/sentrei/commit/3bd12550f6f1a2be250c0497c665e79e9d1ecd88))
+
+## @sentrei/www-mosh-lol [1.5.1](https://github.com/sentrei/sentrei/compare/@sentrei/www-mosh-lol@1.5.0...@sentrei/www-mosh-lol@1.5.1) (2021-09-22)
+
+### Performance Improvements
+
+- ini about me ([5a25d54](https://github.com/sentrei/sentrei/commit/5a25d545b379724bd3301ce3efc7dc22923695b2))
+
 # @sentrei/www-mosh-lol [1.5.0](https://github.com/sentrei/sentrei/compare/@sentrei/www-mosh-lol@1.4.1...@sentrei/www-mosh-lol@1.5.0) (2021-09-21)
 
 ### Features