+++ packages/segment/CHANGELOG.md
@@ -1,3 +1,9 @@
+# @sentrei/segment [1.1.0](https://github.com/sentrei/sentrei/compare/@sentrei/segment@1.0.2...@sentrei/segment@1.1.0) (2021-09-23)
+
+### Features
+
+- ini app mosh lol with json sort ([3bd1255](https://github.com/sentrei/sentrei/commit/3bd12550f6f1a2be250c0497c665e79e9d1ecd88))
+
 ## @sentrei/segment [1.0.2](https://github.com/sentrei/sentrei/compare/@sentrei/segment@1.0.1...@sentrei/segment@1.0.2) (2021-09-20)
 
 ### Performance Improvements